requirements.txt
# Core dependencies
requests>=2.31.0
rich>=13.0.0
pyyaml>=6.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
aiofiles>=23.0.0
jinja2>=3.1.2

# Security & Auth (optional)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Background tasks & caching
redis>=5.0.0
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
